fix end2end training bugs

+++ lib/mod/bbox_nms.py
@@ -137,9 +137,8 @@ class BboxNMSLayer(caffe.Layer):
 
 						#Compensate for batch index
 						boxes_nms_xyxy = np.insert(boxes_nms_xyxy,0,batch_idx,axis=1)
-						
 						blob_rois = np.concatenate((blob_rois, boxes_nms_xyxy))
-						blob_label = np.concatenate((blob_label, np.full((len(boxes_nms_xyxy),1,1,1), labels[batch_idx])))
+						blob_label = np.concatenate((blob_label, np.full((len(boxes_nms_xyxy),1,1,1), labels[batch_idx], dtype=np.float32)))
 
 				else:
 					if len(self.mhi_buf) == 30:
@@ -157,7 +156,7 @@ class BboxNMSLayer(caffe.Layer):
 						pass
 					blob_rois = np.concatenate((blob_rois, np.array([[batch_idx,0,0,128,128],])))
 					blob_label = np.concatenate((blob_label,np.array([[[[0]]]])))
-	
+
 		top[0].reshape(*(blob_rois.shape))
 		top[0].data[...] = blob_rois.astype(np.float32, copy=False)
